읽고 있는 책 내용, 느낀점 디비에 업데이트 기능

+++ app/src/main/java/com/lee/oneweekonebook/database/BookDatabaseDao.kt
@@ -14,7 +14,7 @@ interface BookDatabaseDao {
     fun update(book: Book)
 
     @Query("SELECT * FROM book_history_table WHERE id = :id")
-    fun get(id: Long): Book?
+    fun get(id: Int): Book?
 
     @Query("SELECT * FROM book_history_table")
     fun getAllBooks(): LiveData<List<Book>>
